Added rough tracking of currently sold and analysed biodata.
Status to sold biodata happens to all analysed biodata as soon as any biodata is sold. This version does not account for selling only 1 systems then dying right after and losing the rest of the analysed biodata. To do that the plugin creates and maintians 2 jsonfiles: soldbiodata.json which holds all biodata that was sold. And notsoldbiodata.json which holds only the biodata we analysed and haven't been able to sell yet.

Moved load.py as its removing an unnecessary folder.
As the folder of the plugin does not have to be the exact plugin name.