More granular tracking of biodata and added journal crawler
Added journalcrawler.py, this is a script that goes through any ED player journal and builds a soldbiodata.json.

It lacks checks for keeping it seperate for each player.
This is something I will have to implement or a lot of players
could only use it for a single CMDR.

plugin implementation is untested but the crawler seems to work nicely so far.