now looking at "species" instead of "species_localised"
in journal entries the plugin now looks at the species which is completely independent and always the same for any language given any species.

for this organicinfo got another dictionary to translate from the species' journal name to the in plugins' localised name for a given species.

The journal names are the type that appear in the SellOrganicData

technically this doesn't resolve any opened issues its just here to mitigate any risk of this being the case.