initial multicommander support.
Modified structure of notsoldbiodata.json and soldbiodata.json slightly to allow for multiple commander support.
older jsons are not compatible with this!
Those need to be recreated with the journalcrawling button.

Encountered and fixed a bug with journal crawling where scans of fumerolas couldn't get translated into their english name.