new soldbiodatajson format support for main plugin
Also in this commit:
A fix for an edgecase where the journal crawler and by extension the plugin was adding empty system list pairs to the wrong dictionary letter pair. Since this could also affect the main plugin it was added as a saveguard there aswell.

This should now implement #13

Off to a test and posible debug phase on this feature before I can pack everything together and get the next release out.