fixes for new soldbiodata format.
bugs in biosell event, was missing the replace function that I needed to introduce for the thousands operator.

It's now properly adding the biodata while playing the game.

Had forgotten I do another loop to write the values in the loaded json structure at the end of the biosell event function.

When specifically removing some biodata using the journalcrawling button in the preferences fixes the soldbiodata.json to the state it should be in (given the journals with the required information exist in the first place)

#13 is implemented and functioning now for the new format only.
I still have to check somewhere if the file has the old format used in v0.1.2 or earlier and then react accordingly to it.