Refactor and old format check
plugin now checks if there is already a soldbiodata.json or notsoldbiodata.json in the old v0.1.2 version (main type the json.load returns being still a list and not a dict)

When we encounter the old format we clear the file and treat is as if  the file never existed.

The user then can rebuild the state of the soldbiodata.json once their not yet sold scans are sold to have the full picture about their done exobiology scans and the plugin will take it from there.

This one means I'm finished with the feature #13 on this branch