Pull state for v0.2.1 release to main (#26)
* Implementation of 1st variant

Implemented the first variant I mentioned in #19

Not completely sure I'm content with how the AST_value display behaves in this new mode though I haven't tested it yet.

Additional bugfixes:
- preventing the rare case where cmdr is None or "None" and the build_sold_bio_ui wants to build the sold bio signals.
- catching an error in build_sold_bio_ui where it would go belly up when there was not a current_system yet (before game startup) and not be able to apply a theme in the process

Other:
- shortened save_cmdr
- small changes in preferences ui, shorter lines, using the actual AST_VERSION to display etc.

* Added expand collapse for sold/scanned exobiology

Added a button that expands and collapses the list of sold and analysed exobiology in a system.

Switched all usages of tk.Label to use my own defined ui_label for brevity.

* update to readme and load.py

changed "currentpos" to "Current Position:" in load.py

updated readme.md removed the new tags in front of the features from last release and the info for people that move over from an older version. (the few that used the 0.1.2 should've switched by now anyway, and the info is still in the release notes.)

* small rework of auto hide feature.

now unsold value metric has its own preference tickbox for automatically hiding it when it is 0 Cr.

second way to automatically hide. First way was hiding no longer relevant fields after selling all. new second option is when a scan is fully finished.

* Automatically hide double system/body apperances when there is an auto hide feature on.