Update load.py
changed unsold value tracking to happen in self.rawvalue instead of AST_value that is mainly for displaying it.

Added number shortening function and option to use it via tickbox in the preferences.

When outside of the Clonal Colonial Range the respective entries become green to signify that a new sample can be taken.
added ui_colourlabel and ui_colourentry that just add another parameter to set the text (in tk terminology foreground "fg") colour