Update load.py
forgot that I'm in plugin on the rawvalue is None check.

resurrection also resets scan progress back to 0.

added price for started sampling to the field that shows its name.
new local variables that I needed anyway for this made a few lines in the Analyse case more readable.

moved ui_colourlabel a few lines down.