Implementation of #28
technically it already tracks the "not yet sold data" as "possibly sold data".

Had to add a few extra checks to both the items that got sold (had a duplication issue so that on every rescan of my journals because I changed the method from deleting the old soldbiodata.json and building it anew to updating it)

The same with the notsoldbiodata.json.
Additionally I had elements that were wrongfully added to the notsoldbiodata.json that actually were sold already.
So I needed a specific check for those.