Update load.py
Changes does not impact the way it works on windows which is good.

Have to confirm for linux install of Elite and EDMC.
Might have to set that up one way or the other to confirm locally.

whitespaces before so that flake8 will keep quiet about them missing after the keyword not.