Catching actually expected Error
When we try to use a planet name as a key it can of course be that said planet has no bio signals in this case it is not present in AST_bios_on_planet.

Need to catch the KeyError and let it continue without spamming the debug log, unless we want it to spam the debug log.